% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plots.R
\name{get_plots}
\alias{get_plots}
\title{Plot the Torgegram and the Empirical Semivariogram for a Spatial Stream Network object}
\usage{
get_plots(ssn, ResponseName, Euclidean = FALSE,
  EmpVarMeth = "MethMoment")
}
\arguments{
\item{ssn}{a \link[SSN]{SpatialStreamNetwork-class} object.}

\item{ResponseName}{the name of the response variable.}

\item{Euclidean}{If \code{TRUE} the Empirical Semivariogram based on Euclidean distances is also plotted.}

\item{EmpVarMeth}{a string indicating the name of the method to be used for computing the semivariogram.}
}
\value{
A vector of distance matrices.
}
\description{
Given a \link[SSN]{SpatialStreamNetwork-class} object, computes the Torgegram and the Empirical Semivariogram and provides the plots.
}
\details{
This function works on objects of \link[SSN]{SpatialStreamNetwork-class}.
}
\references{
Peterson, E.E. and Ver Hoef, J.M. (2010) A mixed-model moving-average approach to geostatistical modeling in stream networks. Ecology 91(3), 644–651.

Ver Hoef, J.M. and Peterson, E.E. (2010) A moving average approach for spatial statistical models of stream networks (with discussion). Journal of the American Statistical Association 105, 6–18. DOI: 10.1198/jasa.2009.ap08248. Rejoinder pgs. 22–24.
}
