% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SSN_model.R
\name{get_SSN_model}
\alias{get_SSN_model}
\title{get model}
\usage{
get_SSN_model(varNames, weightVar, CorModels, bin_tables, network_data,
  obs_points, pred_points, obs_data, pred_data)
}
\description{
get model
}
