% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SSN_model_kriging.R
\name{get_SSN_model_kriging}
\alias{get_SSN_model_kriging}
\title{get model and kriging}
\usage{
get_SSN_model_kriging(ssn, varNames, weightVar, predpts, CorModels,
  useNugget = TRUE)
}
\description{
get model and kriging
}
