% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ssn.R
\name{plot.ssn}
\alias{plot.ssn}
\title{Plot the predicted values along a Spatial Stream Network object}
\usage{
plot.ssn(x, VariableName = NULL, color.palette = NULL,
  nclasses = NULL, breaktype = "quantile", brks = NULL, lb = NULL,
  ub = NULL, PredPointsID = NULL, add = FALSE,
  addWithLegend = FALSE, lwdLineCol = NULL, lwdLineEx = 1,
  lineCol = "black", ...)
}
\arguments{
\item{VariableName}{name of variable to be plotted, it can be also a factor.}

\item{color.palette}{a color palette for plotting points. The default is rainbow(nclasses, start = .66, end = .99). The number of colors should equal to the number of classes. See \code{\link{palette}} for many ways to create palettes.}

\item{nclasses}{the number of classes for coloring the predictions (or standard errors) according to their value.  The default is 10. If brks = c(...) is specified, then nclasses is automatically set to the number of breaks + 1.}

\item{breaktype}{the method for breaking the predictions (or standard errors) into classes for coloring while plotting. A character argument that must be one of "quantile" (default), "even", or "user".}

\item{brks}{if breaktype = "user", the break values must be specified here as a vector or matrix using c(...) or cbind(...). The sorted unique values are used as break points (together with the min and max of the variable being plotted if required)}

\item{lb}{minimum value for the break values, otherwise computed as the minimum value according to the chosen method for breaking the predictions.}

\item{ub}{maximum value for the break values, otherwise computed as the maximum value according to the chosen method for breaking the predictions.}

\item{PredPointsID}{a string representing the internal name of the prediction sites data set, which will be added to the plot. Default is NULL.}

\item{add}{Logical value indicating whether the predictions should be added to an existing plot, such as a plot of colored values for observed data. Default is FALSE.}

\item{addWithLegend}{logical indicating whether the predictions should be added to an existing plot, such as a plot of colored values for observed data. Use this when there is a legend. Default is FALSE.}

\item{lwdLineCol}{a column name in the lines data frame to be used for line width expansion. This will most likely be the name of the additive function column, but others could be used.}

\item{lwdLineEx}{an expansion multiplier to create line widths for the values contained in lwdLineCol.}

\item{lineCol}{a color for the lines forming the stream network. Default is "black".}

\item{\dots}{Arguments to be passed to methods, such as graphical parameters (see \code{\link{par}}).}

\item{ssn.object}{a \link[SSN]{SpatialStreamNetwork-class} object.}
}
\value{
MMaps of stream networks.
}
\description{
Plot the predicted values along a Spatial Stream Network object
}
\details{
The plot.ssn function creates a map showing data locations that can be color-coded according to the values of observed variables. Prediction locations can also be added to existing plots of observed values.
}
\references{
Peterson, E.E. and Ver Hoef, J.M. (2010) A mixed-model moving-average approach to geostatistical modeling in stream networks. Ecology 91(3), 644–651.

Ver Hoef, J.M. and Peterson, E.E. (2010) A moving average approach for spatial statistical models of stream networks (with discussion). Journal of the American Statistical Association 105, 6–18.
}
