% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Torg.R
\name{plot.Torg}
\alias{plot.Torg}
\title{Plotting Method for Torgegram Objects}
\usage{
plot.Torg(x, sp.relationship = c("fc", "fu"), min.cex = 1.5,
  max.cex = 6, leg.auto = TRUE, main = "", ylab = "",
  xlab = "Stream Distance")
}
\arguments{
\item{x}{an object of class \link[SSN]{Torgegram}.}

\item{sp.relationship}{a string or character vector representing the in-stream spatial relationship to be plotted. "fc" specifies plotting of only flow-connected, and "fu" specifies plotting of only flow-unconnected. Default is both.}

\item{min.cex}{Minimum character expansion size of the plotting symbols. Symbols are scaled according to how many pairs of points went into computing each bin of the semivariogram. The bin with the smallest sample size will be be plotted with this symbol size. The default is 1.5.}

\item{max.cex}{Maximum character expansion size of the plotting symbols. Symbols are scaled according to how many pairs of points went into computing each bin of the semivariogram. The bin with the largest sample size will be be plotted with this symbol size. The default is 6.}

\item{leg.auto}{Logical, default to TRUE. Include a legend.}

\item{main}{Title for plot.}

\item{ylab}{Label for y-axis.}

\item{xlab}{Label for x-axis.}
}
\value{
Plot of empirical semivariogram values
}
\description{
\code{plot}.Torgegram is a generic plot function that has been adapted for Torgegram objects, which are created using the Torgegram function. A Torgegram object stores information used to construct an empirical semivariogram based on hydrologic distance. The \code{plot.Torgegram} function allows the results to be presented separately for flow-connected and flow-unconnected sites.
}
\details{
The \code{Torgegram} function creates a list of distances and empirical semivariogram values, along with number of pairs of points in each bin, for both flow-connected and flow-unconnected sites. Flow-connected locations lie on the same stream network (share a common downstream junction) and water flows from one location to the other. Flow-unconnected locations also lie on the same stream network, but do not share flow. The output is of class Torgegram. This is the default plotting method for this class.
}
\references{
Peterson, E.E. and Ver Hoef, J.M. (2010) A mixed-model moving-average approach to geostatistical modeling in stream networks. Ecology 91(3), 644–651.

Ver Hoef, J.M. and Peterson, E.E. (2010) A moving average approach for spatial statistical models of stream networks (with discussion). Journal of the American Statistical Association 105, 6–18. DOI: 10.1198/jasa.2009.ap08248. Rejoinder pgs. 22–24.
}
