% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_SSN_kriging.R
\name{do_SSN_kriging}
\alias{do_SSN_kriging}
\title{Make prediction on a Spatial Stream Network object}
\usage{
do_SSN_kriging(ssn, varNames, weightVar, predpts, CorModels,
  useNugget = TRUE, theta, covMat, singleNet = NULL, matrices = NULL)
}
\arguments{
\item{ssn}{a \link[SSN]{SpatialStreamNetwork-class} object.}

\item{varNames}{a vector of strings, the names of the variables used in the model.}

\item{weightVar}{a string indicating the name of the variable to compute the spatial weights.}

\item{predpts}{a string indicating the name of the group of points in which make predictions.}

\item{CorModels}{a vector of strings, the names of the covariance models.}

\item{useNugget}{If \code{FALSE} the nugget effect is not included in the model. Default to \code{TRUE}.}

\item{theta}{vector of the parameters values used for computing the covariance matrix.}

\item{covMat}{covariance matrix of the observed points.}

\item{singleNet}{an interger, indicating the network ID that is to be analysed. Default to \code{NULL}, so that the analysis is carried on the entire dataset.}

\item{matrices}{a vector of matrices, containing the flow-connection binary matrix, the hydrological distance matrix and, not necessarily, the Euclidean distance matrix, returned by the function \link[StreamNetgstat]{get_plots}. These matrices consider the relationships between observed points.}
}
\value{
A \code{ssn.object}{ (updated \link[SSN]{SpatialStreamNetwork-class} object), with predicted values included }
}
\description{
Given a \link[SSN]{SpatialStreamNetwork-class} object and the covariance parameters values, makes predictions using universal kriging.
}
\details{
This function calculates prediction values and kriging variance for prediction sites based on the results of a linear model for the \link[SSN]{SpatialStreamNetwork-class} object.
}
\references{
Peterson, E.E. and Ver Hoef, J.M. (2010) A mixed-model moving-average approach to geostatistical modeling in stream networks. Ecology 91(3), 644–651.

Ver Hoef, J.M. and Peterson, E.E. (2010) A moving average approach for spatial statistical models of stream networks (with discussion). Journal of the American Statistical Association 105, 6–18.
}
