% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SSN_model_kriging.R
\name{get_SSN_model_kriging}
\alias{get_SSN_model_kriging}
\title{Fit a variance component model for a Spatial Stream Network object 
and make prediction}
\usage{
get_SSN_model_kriging(ssn, varNames, weightVar, predpts, CorModels,
  useNugget = TRUE)
}
\arguments{
\item{ssn}{a \link[SSN]{SpatialStreamNetwork-class} object.}

\item{varNames}{a vector of strings, the names of the variables used in the model.}

\item{weightVar}{a string indicating the name of the variable to compute the spatial weights.}

\item{predpts}{a string indicating the name of the group of points in which make predictions.}

\item{CorModels}{a vector of strings, the names of the covariance models.}

\item{useNugget}{If \code{FALSE} the nugget effect is not included in the model. Default to \code{TRUE}.}
}
\value{
A list with the following fields:
\item{\code{optTheta}}{ vector of the parameters values of the fitted model. }
\item{\code{betaValues}}{ vector of the beta values of the fitted model. }
\item{\code{covMatrix}}{ covariance matrix of the fitted model. }
\item{\code{predictions}}{ 2-column matrix containing the predicted values and the kriging variance. }
}
\description{
Given a \link[SSN]{SpatialStreamNetwork-class} object, fits a linear model using a variance component approach. Then it makes predictions using universal kriging.
}
\details{
This function works on objects of \link[SSN]{SpatialStreamNetwork-class} to fit generalized linear models with spatially autocorrelated errors using normal likelihood methods. Then it calculates prediction values and kriging variance for prediction sites.
}
\references{
Peterson, E.E. and Ver Hoef, J.M. (2010) A mixed-model moving-average approach to geostatistical modeling in stream networks. Ecology 91(3), 644–651.

Ver Hoef, J.M. and Peterson, E.E. (2010) A moving average approach for spatial statistical models of stream networks (with discussion). Journal of the American Statistical Association 105, 6–18. DOI: 10.1198/jasa.2009.ap08248. Rejoinder pgs. 22–24.
}
