% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SSN_model_kriging.R
\name{get_SSN_model_kriging}
\alias{get_SSN_model_kriging}
\title{Fit a variance component model for a Spatial Stream Network object 
and make prediction}
\usage{
get_SSN_model_kriging(ssn, varNames, weightVar, predpts, CorModels,
  useNugget = TRUE, singleNet = NULL, matrices = NULL,
  bounds = NULL, useCholeskyDec = FALSE)
}
\arguments{
\item{ssn}{a \link[SSN]{SpatialStreamNetwork-class} object.}

\item{varNames}{a vector of strings, the names of the variables used in the model.}

\item{weightVar}{a string indicating the name of the variable to compute the spatial weights.}

\item{predpts}{a string indicating the name of the group of points in which make predictions.}

\item{CorModels}{a vector of strings, the names of the covariance models. The individual models should be of different "types" tail-up, tail-down, Euclidean. The tailup models include: "Exponential.tailup", "LinearSill.tailup", "Spherical.tailup", 

"Mariah.tailup" "Epanech.tailup"; tail-down models include: "Exponential.taildown", "LinearSill.taildown", "Spherical.taildown", "Mariah.taildown"; Euclidean distance models include: "Spherical.Euclid", "Gaussian.Euclid", "Exponential.Euclid", 

"Cauchy.Euclid". The first 4 tailup and taildown models are described in Ver Hoef and Peterson (2010), and the 4 Euclidean distance models are standard spatial covariance models.}

\item{useNugget}{If \code{FALSE} the nugget effect is not included in the model. Default to \code{TRUE}.}

\item{singleNet}{an interger, indicating the network ID that is to be analysed. Default to \code{NULL}, so that the analysis is carried on the entire dataset.}

\item{matrices}{a vector of matrices, containing the flow-connection binary matrix, the hydrological distance matrix and, not necessarily, the Euclidean distance matrix, returned by the function \link[StreamNetgstat]{get_plots}. These matrices consider the relationships between observed points.}

\item{bounds}{a vector of doubles, representing the bounds for the parsills of the models considered. If a bound is required, all the models should have one. The highest can be set at 1e+04.}

\item{useCholeskyDec}{If \code{TRUE} the Cholesky decomposition for inverting positive definite matrices is always used in the optimization algorithm (quickest version, at the expense of accuracy). Default to \code{FALSE}.}
}
\value{
A list with the following fields:
\item{\code{ssn.object}}{ updated \link[SSN]{SpatialStreamNetwork-class} object, with predicted values included. }
\item{\code{modelParam}}{ vector of the parameters values of the fitted model. }
\item{\code{modelBeta}}{ vector of the beta values of the fitted model. }
}
\description{
Given a \link[SSN]{SpatialStreamNetwork-class} object, fits a linear model using a variance component approach. Then it makes predictions using universal kriging.
}
\details{
This function works on objects of \link[SSN]{SpatialStreamNetwork-class} to fit generalized linear models with spatially autocorrelated errors using normal likelihood methods. Then it calculates prediction values and kriging variance for prediction sites.
}
\references{
Peterson, E.E. and Ver Hoef, J.M. (2010) A mixed-model moving-average approach to geostatistical modeling in stream networks. Ecology 91(3), 644–651.

Ver Hoef, J.M. and Peterson, E.E. (2010) A moving average approach for spatial statistical models of stream networks (with discussion). Journal of the American Statistical Association 105, 6–18.
}
