% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ssn.R
\name{plot.ssn}
\alias{plot.ssn}
\title{Plot the predicted values along a Spatial Stream Network object}
\usage{
plot.ssn(x, VariableName = NULL, color.palette = NULL,
  nclasses = NULL, breaktype = "quantile", brks = NULL,
  PredPointsID = NULL, add = FALSE, addWithLegend = FALSE,
  lwdLineCol = NULL, lwdLineEx = 1, lineCol = "black", ...)
}
\arguments{
\item{VariableName}{name of variable to be plotted, it can be also a factor.}

\item{color.palette}{a color palette for plotting points. The default is rainbow(nclasses, start = .66, end = .99). The number of colors should equal to the number of classes. See \code{\link{palette}} for many ways to create palettes.}

\item{nclasses}{the number of classes for coloring the predictions (or standard errors) according to their value.  The default is 10. If brks = c(...) is specified, then nclasses is automatically set to the number of breaks + 1.}

\item{breaktype}{the method for breaking the predictions (or standard errors) into classes for coloring while plotting. A character argument that must be one of "quantile" (default), "even", or "user".}

\item{brks}{if breaktype = "user", the break values must be specified here as a vector or matrix using c(...) or cbind(...). The sorted unique values are used as break points (together with the min and max of the variable being plotted if required)}

\item{add}{Logical value indicating whether the predictions should be added to an existing plot, such as a plot of colored values for observed data. Default is FALSE.}

\item{\dots}{Arguments to be passed to methods, such as graphical parameters (see \code{\link{par}}).}

\item{ssn.object}{a \link[SSN]{SpatialStreamNetwork-class} object.}

\item{VarPlot}{a character argument that must be one of "Both", "Predictions", or "Standard Errors". Default is "Both", which colors predictions by their values and makes their size inversely proportional to the prediction standard errors.}

\item{SEcex.min}{if VarPlot = "both", the minimum cex value when making point sizes is inversely proportional to the prediction standard errors. See \code{\link{par}} for more on cex.  Also see details below. Default is 1.}

\item{SEcex.max}{if VarPlot = "both", the maximum cex value when making point sizes inversely proportional to the prediction standard errors. See \code{\link{par}} for more on cex.  Also see details below. Default is 3.}

\item{dec.dig}{the number of decimal places to print in the legend.  Default is 2.}
}
\value{
Maps of stream networks with prediction and prediction standard error values.
}
\description{
Plot the predicted values along a Spatial Stream Network object
}
\details{
The \command{plot.predictions} function creates a map showing color-coded predictions or prediction standard error values. When VarPlot = "Both", predictions values are colored according to breaks.  The size of the points is inversely proportional to the prediction standard errors.
}
\references{
Peterson, E.E. and Ver Hoef, J.M. (2010) A mixed-model moving-average approach to geostatistical modeling in stream networks. Ecology 91(3), 644–651.

Ver Hoef, J.M. and Peterson, E.E. (2010) A moving average approach for spatial statistical models of stream networks (with discussion). Journal of the American Statistical Association 105, 6–18.
}
