% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.torg.R
\name{plot.torg}
\alias{plot.torg}
\title{Plotting Method for Torgegram Objects}
\usage{
plot.torg(x, sp.relationship = c("fc", "fu"), min.cex = 1.5,
  max.cex = 6, leg.auto = TRUE, main = "", ylab = "",
  xlab = "Stream Distance")
}
\arguments{
\item{x}{an object of class \link[SSN]{Torgegram}.}

\item{sp.relationship}{a string or character vector representing the in-stream spatial relationship to be plotted. "fc" specifies plotting of only flow-connected, and "fu" specifies plotting of only flow-unconnected. Default is both.}

\item{min.cex}{Minimum character expansion size of the plotting symbols. Symbols are scaled according to how many pairs of points went into computing each bin of the semivariogram. The bin with the smallest sample size will be be plotted with this symbol size. The default is 1.5.}

\item{max.cex}{Maximum character expansion size of the plotting symbols. Symbols are scaled according to how many pairs of points went into computing each bin of the semivariogram. The bin with the largest sample size will be be plotted with this symbol size. The default is 6.}

\item{leg.auto}{Logical, default to TRUE. Include a legend.}

\item{main}{Title for plot.}

\item{ylab}{Label for y-axis.}

\item{xlab}{Label for x-axis.}
}
\value{
Plot of empirical semivariogram values.
}
\description{
\code{plot}.Torg is a generic plot function that has been adapted for \link[SSN]{Torgegram} objects, which are created using the \link[StreamNetgstat]{torgegram} function.
}
\details{
This function provides the plot of the empirical semivariogram based on stream distance. The results are presented separately for flow-connected and flow-unconnected sites.
}
\references{
Peterson, E.E. and Ver Hoef, J.M. (2010) A mixed-model moving-average approach to geostatistical modeling in stream networks. Ecology 91(3), 644–651.

Ver Hoef, J.M. and Peterson, E.E. (2010) A moving average approach for spatial statistical models of stream networks (with discussion). Journal of the American Statistical Association 105, 6–18.
}
